flask_demo/pyproject.toml
[project]
name = "flask-caching-demo"
version = "0.1.0"
description = "Flask web application demonstrating best practices for database and API caching using Valkey (Redis-compatible)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Flask>=3.0.0",
    "Flask-SQLAlchemy>=3.1.0",
    "Flask-Caching>=2.1.0",
    "Flask-DebugToolbar>=0.15.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0"
]
